support Prestashop v9

+++ controllers/admin/AdminCrispController.php
@@ -9,7 +9,7 @@
  * You must not modify, adapt or create derivative works of this source code
  *
  *  @author    Crisp IM SAS
- *  @copyright 2024 Crisp IM SAS
+ *  @copyright 2026 Crisp IM SAS
  *  @license   All rights reserved to Crisp IM SAS
  */
 if (!defined('_PS_VERSION_')) {
@@ -17,7 +17,6 @@ if (!defined('_PS_VERSION_')) {
 }
 
 use PrestaShop\PrestaShop\Core\Addon\Module\ModuleManagerBuilder;
-use PrestaShop\PsAccountsInstaller\Installer\Exception\InstallerException;
 
 class AdminCrispController extends ModuleAdminController
 {
@@ -25,8 +24,6 @@ class AdminCrispController extends ModuleAdminController
 
     public function __construct()
     {
-        $crisp = Module::getInstanceByName('crisp');
-        $this->module = $crisp;
         $this->bootstrap = true;
         $this->table = 'configuration';
         $this->className = 'Configuration';
@@ -34,26 +31,33 @@ class AdminCrispController extends ModuleAdminController
         $this->display = 'view';
 
         parent::__construct();
+
+        $crisp = Module::getInstanceByName('crisp');
+        if ($crisp instanceof Crisp) {
+            $this->module = $crisp;
+        }
     }
 
-    public function displayAjaxEnableChatbox()
+    public function displayAjaxEnableChatbox(): void
     {
         Configuration::updateValue('CRISP_CHATBOX_DISABLED', 0);
     }
 
-    public function displayAjaxDisableChatbox()
+    public function displayAjaxDisableChatbox(): void
     {
         Configuration::updateValue('CRISP_CHATBOX_DISABLED', 1);
     }
 
-    public function displayAjaxEnableWebService()
+    public function displayAjaxEnableWebService(): void
     {
         // Delete any Webservice keys created by Crisp.
         $crisp_webservice_key_id = Configuration::get('CRISP_WEBSERVICE_KEY_ID');
 
-        $webserviceKey = new WebserviceKey($crisp_webservice_key_id);
-        if (Validate::isLoadedObject($webserviceKey)) {
-            $webserviceKey->delete();
+        if ($crisp_webservice_key_id) {
+            $webserviceKey = new WebserviceKey((int) $crisp_webservice_key_id);
+            if (Validate::isLoadedObject($webserviceKey)) {
+                $webserviceKey->delete();
+            }
         }
 
         // Fetch new API key from URL
@@ -83,37 +87,41 @@ class AdminCrispController extends ModuleAdminController
         WebserviceKey::setPermissionForAccount($apiAccess->id, $permissions);
     }
 
-    public function displayAjaxToggleEnableWebservice()
+    public function displayAjaxToggleEnableWebservice(): void
+    {
+        // Enable any Webservice keys created by Crisp.
+        $crisp_webservice_key_id = Configuration::get('CRISP_WEBSERVICE_KEY_ID');
+
+        if ($crisp_webservice_key_id) {
+            $webserviceKey = new WebserviceKey((int) $crisp_webservice_key_id);
+            if (Validate::isLoadedObject($webserviceKey)) {
+                $webserviceKey->active = true;
+                $webserviceKey->save();
+            }
+        }
+    }
+
+    public function displayAjaxToggleDisableWebservice(): void
     {
         // Disable any Webservice keys created by Crisp.
         $crisp_webservice_key_id = Configuration::get('CRISP_WEBSERVICE_KEY_ID');
 
-        $webserviceKey = new WebserviceKey($crisp_webservice_key_id);
-        if (Validate::isLoadedObject($webserviceKey)) {
-            $webserviceKey->active = true;
-            $webserviceKey->save();
+        if ($crisp_webservice_key_id) {
+            $webserviceKey = new WebserviceKey((int) $crisp_webservice_key_id);
+            if (Validate::isLoadedObject($webserviceKey)) {
+                $webserviceKey->active = false;
+                $webserviceKey->save();
+            }
         }
     }
 
-    public function displayAjaxToggleDisableWebservice()
-    {
-        // Delete any Webservice keys created by Crisp.
-        $crisp_webservice_key_id = Configuration::get('CRISP_WEBSERVICE_KEY_ID');
-
-        $webserviceKey = new WebserviceKey($crisp_webservice_key_id);
-        if (Validate::isLoadedObject($webserviceKey)) {
-            $webserviceKey->active = false;
-            $webserviceKey->save();
-        }
-    }
-
-    public function renderView()
+    public function renderView(): string
     {
         $link = new Link();
         $admin_url = $link->getAdminLink('AdminCrisp');
         $chatbox_disabled = Configuration::get('CRISP_CHATBOX_DISABLED');
 
-        $http_callback = 'http' . (($_SERVER['SERVER_PORT'] == 443) ? 's://' : '://') . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
+        $http_callback = Tools::getShopProtocol() . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
         $crisp_installed = false;
         $get_website_id = Tools::getValue('crisp_website_id');
         if (isset($get_website_id) && !empty($get_website_id)) {
@@ -133,16 +141,18 @@ class AdminCrispController extends ModuleAdminController
                 $api_key_invalid = true;
             }
 
-            $webserviceKey = new WebserviceKey($crisp_webservice_key_id);
-            // Consider api key disabled if new API key does not match existing API key.
-            if (Validate::isLoadedObject($webserviceKey) && $webserviceKey->key !== $crisp_api_key) {
-                $api_key_disabled = true;
-                $api_key_invalid = true;
+            if ($crisp_webservice_key_id) {
+                $webserviceKey = new WebserviceKey((int) $crisp_webservice_key_id);
+                // Consider api key disabled if new API key does not match existing API key.
+                if (Validate::isLoadedObject($webserviceKey) && $webserviceKey->key !== $crisp_api_key) {
+                    $api_key_disabled = true;
+                    $api_key_invalid = true;
+                }
             }
         }
 
         if (!empty($crisp_webservice_key_id)) {
-            $webserviceKey = new WebserviceKey($crisp_webservice_key_id);
+            $webserviceKey = new WebserviceKey((int) $crisp_webservice_key_id);
 
             if (Validate::isLoadedObject($webserviceKey) && !$webserviceKey->active) {
                 $api_key_disabled = true;
@@ -151,16 +161,15 @@ class AdminCrispController extends ModuleAdminController
 
         $admin_locale = $this->context->language->iso_code;
 
-        $context = Context::getContext();
-        $shop = $context->shop;
+        $shop = $this->context->shop;
         $shopName = Configuration::get('PS_SHOP_NAME');
         $shopDomain = $shop->domain;
         $adminEmail = '';
         $adminUsername = '';
 
-        if (isset($context->employee) && $context->employee->id) {
-            $adminUsername = $context->employee->firstname . ' ' . $context->employee->lastname;
-            $adminEmail = $context->employee->email;
+        if (isset($this->context->employee) && $this->context->employee->id) {
+            $adminUsername = $this->context->employee->firstname . ' ' . $this->context->employee->lastname;
+            $adminEmail = $this->context->employee->email;
         }
 
         $website_id = Configuration::get('WEBSITE_ID');
@@ -179,6 +188,7 @@ class AdminCrispController extends ModuleAdminController
             'crisp_webservice_key_id' => $crisp_webservice_key_id,
             'crisp_plugin_identifier' => CRISP_PLUGIN_IDENTIFIER,
             'crisp_plugin_url' => CRISP_PLUGIN_URL,
+            'crisp_app_url' => CRISP_APP_URL,
             'crisp_plugin_source' => CRISP_PLUGIN_SOURCE,
             'logo' => '/modules/crisp/logo.png',
             'shop_name' => $shopName,
@@ -189,51 +199,66 @@ class AdminCrispController extends ModuleAdminController
 
         $this->context->controller->addCSS(_PS_MODULE_DIR_ . 'crisp/views/css/style.css', 'all');
         $this->context->controller->addCSS('https://ui-kit.prestashop.com/backoffice/latest/css/bootstrap-prestashop-ui-kit.css', 'all');
-        $this->context->controller->addJS('https://ui-kit.prestashop.com/backoffice/latest/js/prestashop-ui-kit.js', 'all');
-        $this->context->controller->addJS(_PS_MODULE_DIR_ . 'crisp/views/js/petite-vue.js', 'all');
-        $this->context->controller->addJS(_PS_MODULE_DIR_ . 'crisp/views/js/admin.js', 'all');
+        $this->context->controller->addJS('https://ui-kit.prestashop.com/backoffice/latest/js/prestashop-ui-kit.js');
+        $this->context->controller->addJS(_PS_MODULE_DIR_ . 'crisp/views/js/petite-vue.js');
+        $this->context->controller->addJS(_PS_MODULE_DIR_ . 'crisp/views/js/admin.js');
+
         $moduleManager = ModuleManagerBuilder::getInstance()->build();
-        $accountsService = null;
-        try {
-            $accountsFacade = $this->module->getService('crisp.ps_accounts_facade');
-            $accountsService = $accountsFacade->getPsAccountsService();
-        } catch (InstallerException $e) {
-            $accountsInstaller = $this->module->getService('crisp.ps_accounts_installer');
-            $accountsInstaller->install();
-            $accountsFacade = $this->module->getService('crisp.ps_accounts_facade');
-            $accountsService = $accountsFacade->getPsAccountsService();
-        }
-        try {
-            Media::addJsDef([
-                'contextPsAccounts' => $accountsFacade->getPsAccountsPresenter()->present($this->module->name),
-            ]);
 
-            $this->context->smarty->assign('urlAccountsCdn', $accountsService->getAccountsCdn());
-        } catch (Exception $e) {
-            $this->context->controller->errors[] = $e->getMessage();
+        $urlAccountsCdn = '';
+        $urlCloudsync = '';
 
-            return '';
-        }
-        if ($moduleManager->isInstalled('ps_eventbus')) {
-            $eventbusModule = Module::getInstanceByName('ps_eventbus');
-            $eventbusScope = [
-                'info',
-                'customers',
-                'orders',
-                'carriers',
-                'carts',
-                'currencies',
-                'products',
-            ];
+        if ($moduleManager->isInstalled('ps_accounts')) {
+            $psAccountsModule = Module::getInstanceByName('ps_accounts');
+            if ($psAccountsModule !== false && method_exists($psAccountsModule, 'getService')) {
+                try {
+                    $accountsService = $psAccountsModule->getService('PrestaShop\Module\PsAccounts\Service\PsAccountsService');
+                    $accountsPresenter = $psAccountsModule->getService('PrestaShop\Module\PsAccounts\Presenter\PsAccountsPresenter');
 
-            if (version_compare($eventbusModule->version, '1.9.0', '>=')) {
-                $eventbusPresenterService = $eventbusModule->getService('PrestaShop\Module\PsEventbus\Service\PresenterService');
-                $this->context->smarty->assign('urlCloudsync', 'https://assets.prestashop3.com/ext/cloudsync-merchant-sync-consent/latest/cloudsync-cdc.js');
-                Media::addJsDef([
-                    'contextPsEventbus' => $eventbusPresenterService->expose($this->module, $eventbusScope),
-                ]);
+                    if (method_exists($accountsPresenter, 'present') && method_exists($accountsService, 'getAccountsCdn')) {
+                        Media::addJsDef([
+                            'contextPsAccounts' => $accountsPresenter->present('crisp'),
+                        ]);
+                        $urlAccountsCdn = $accountsService->getAccountsCdn();
+                    }
+                } catch (\Exception $e) {
+                    $this->context->controller->errors[] = 'PS Accounts: ' . $e->getMessage();
+                }
             }
         }
+
+        if ($moduleManager->isInstalled('ps_eventbus')) {
+            $eventbusModule = Module::getInstanceByName('ps_eventbus');
+            if ($eventbusModule !== false && method_exists($eventbusModule, 'getService')) {
+                $eventbusScope = [
+                    'info',
+                    'customers',
+                    'orders',
+                    'carriers',
+                    'carts',
+                    'currencies',
+                    'products',
+                ];
+
+                if (version_compare($eventbusModule->version, '1.9.0', '>=')) {
+                    try {
+                        $eventbusPresenterService = $eventbusModule->getService('PrestaShop\Module\PsEventbus\Service\PresenterService');
+                        if (method_exists($eventbusPresenterService, 'expose')) {
+                            $urlCloudsync = 'https://assets.prestashop3.com/ext/cloudsync-merchant-sync-consent/latest/cloudsync-cdc.js';
+                            Media::addJsDef([
+                                'contextPsEventbus' => $eventbusPresenterService->expose($this->module, $eventbusScope),
+                            ]);
+                        }
+                    } catch (\Exception $e) {
+                        $this->context->controller->errors[] = $e->getMessage();
+                    }
+                }
+            }
+        }
+
+        $this->context->smarty->assign('urlAccountsCdn', $urlAccountsCdn);
+        $this->context->smarty->assign('urlCloudsync', $urlCloudsync);
+
         $moduleTemplatePath = _PS_MODULE_DIR_ . 'crisp/views/templates/admin/configure.tpl';
         $templateContent = $this->context->smarty->fetch($moduleTemplatePath);
